third_party/blink/web_tests/http/tests/preload/meta-viewport-link-headers-imagesrcset.html

<!DOCTYPE html>
<script>
    if (window.testRunner) {
        internals.settings.setViewportEnabled(true);
        internals.settings.setViewportMetaEnabled(true);
        testRunner.setPopupBlockingEnabled(false);
    }
</script>
<meta name="viewport" content="width=160">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that Link headers support the imagesrcset and imagesizes attributes and respond to <meta content=viewport>');
    addEventListener("message", t.step_func(function(event) {
        assert_array_equals(event.data.sort(), ["300", "small"])
        t.done();
    }, false));
    window.open("resources/enable-viewport-and-refresh.html?media-link-headers-imagesrcset.php");
</script>
